Movement based on mouse click
Fixed up the grid based movement so that units will now move through the center of the squares instead of along the edges. Also fixed an off by one bug in the MoveAlongPath() method that caused a unit to stop moving 1 square before its destination. Lastly, added an update check for the unit where if there is a mouse click it runs A* pathfinding to see if there is a path to the square clicked. If so, it moves the unit to that square smoothly. This will need to be changed around quite a bit to roll it properly into the turn-based system but for now it works well for testing how well the movement works.